# api/internal/runtime/static/openapi.yaml
openapi: 3.0.3
info:
  title: NAAP Analytics API
  version: "1"
  description: |
    Read-only analytics API for the Livepeer AI network.
    Backed by ClickHouse materialized views fed from two Kafka topics:
    `network_events` (daydream org) and `streaming_events` (cloudspe org).

    **Rate limiting:** 100 req/s sustained, burst 200 per IP.
    Exceeded requests receive `429 Too Many Requests`.

    **Default time window:** last 24 hours when `start`/`end` are omitted.

    **WEI values** (payment amounts) are returned as decimal strings to avoid
    float64 precision loss (e.g. `"1234500000000000000"`).
servers:
  - url: https://naap-api.cloudspe.com
    description: Production
  - url: http://localhost:8000
    description: Local development
  - url: https://naap-api.livepeer.cloud
    description: Production
...
components:
  parameters:
    org:
      name: org
      in: query
      description: Filter to a single organisation. Omit for all orgs.
      schema:
        type: string
        enum: [daydream, cloudspe]
    start:
      name: start
      in: query
      description: Window start (RFC3339). Defaults to 24 hours ago.
      schema:
        type: string
        format: date-time
        example: "2026-03-01T00:00:00Z"
    end:
      name: end
      in: query
      description: Window end (RFC3339). Defaults to now.
      schema:
        type: string
        format: date-time
        example: "2026-03-24T23:59:59Z"
    limit:
      name: limit
      in: query
      description: Maximum rows to return (1–1000).
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
    offset:
      name: offset
      in: query
      description: Pagination offset.
      schema:
        type: integer
        minimum: 0
        default: 0
    pipeline:
      name: pipeline
      in: query
      description: Filter by pipeline name (e.g. `text-to-image`).
      schema:
        type: string
    orch_address:
      name: orch_address
      in: query
      description: Filter by orchestrator Ethereum address (lowercased).
      schema:
        type: string
        example: "0xabc123def456"
...
    stream_id:
      name: stream_id
      in: query
      schema:
        type: string
    failure_type:
      name: failure_type
      in: query
      description: Filter failure events by type.
      schema:
        type: string
        enum: [no_orch_available, orch_swap, inference_restart, inference_error]
    address_path:
      name: address
      in: path
      required: true
      description: Orchestrator Ethereum address (lowercased).
      schema:
        type: string
        example: "0xabc123def456"
  schemas:
...
    Problem:
      description: RFC 7807 Problem Details error response.
      type: object
      properties:
        type:
          type: string
          example: "about:blank"
        title:
          type: string
          example: "Too Many Requests"
        status:
          type: integer
          example: 429
        detail:
          type: string
          example: "rate limit exceeded"
...
    PipelineFPS:
      type: object
      properties:
        Pipeline:
          type: string
        InferenceFPS:
          $ref: "#/components/schemas/FPSStats"
        InputFPS:
          $ref: "#/components/schemas/FPSStats"
        SampleCount:
          type: integer
          format: int64
    OrchFPS:
      type: object
      properties:
        Address:
          type: string
        Name:
          type: string
        Pipeline:
          type: string
        InferenceFPS:
          $ref: "#/components/schemas/FPSStats"
        InputFPS:
          $ref: "#/components/schemas/FPSStats"
        SampleCount:
          type: integer
          format: int64
    FPSSummary:
      type: object
      properties:
        StartTime:
          type: string
          format: date-time
        EndTime:
          type: string
          format: date-time
        ByPipeline:
          type: array
          items:
            $ref: "#/components/schemas/PipelineFPS"
        ByOrchestrator:
          type: array
          items:
            $ref: "#/components/schemas/OrchFPS"
    FPSBucket:
      type: object
      properties:
        Timestamp:
          type: string
          format: date-time
        AvgInferenceFPS:
          type: number
          format: double
        AvgInputFPS:
          type: number
          format: double
        SampleCount:
          type: integer
          format: int64
        Degraded:
          type: boolean
          description: True when avg inference FPS drops ≥20% from the prior bucket.
    OrchLatency:
      type: object
      properties:
        Address:
          type: string
        Name:
          type: string
        AvgMS:
          type: number
          format: double
        P50MS:
          type: number
          format: double
        P95MS:
          type: number
          format: double
        P99MS:
          type: number
          format: double
        SampleCount:
          type: integer
          format: int64
    LatencySummary:
      type: object
      properties:
        NetworkAvgMS:
          type: number
          format: double
        ByOrchestrator:
          type: array
          items:
            $ref: "#/components/schemas/OrchLatency"
    WebRTCQuality:
      type: object
      properties:
        StartTime:
          type: string
          format: date-time
        EndTime:
          type: string
          format: date-time
        Video:
          type: object
          properties:
            AvgJitterMS:
              type: number
              format: double
            PacketLossRate:
              type: number
              format: double
        Audio:
          type: object
          properties:
            AvgJitterMS:
              type: number
              format: double
            PacketLossRate:
              type: number
              format: double
        ConnQuality:
          type: object
          properties:
            Good:
              type: number
              format: double
              description: Fraction of streams in good quality tier (0–1).
            Fair:
              type: number
              format: double
            Poor:
              type: number
              format: double

    # ── Payments (R4) ───────────────────────────────────────────────────────
...
    FailureBreakdown:
      type: object
      properties:
        NoOrchAvailable:
          type: integer
          format: int64
        OrchSwap:
          type: integer
          format: int64
        InferenceRestart:
          type: integer
          format: int64
        InferenceError:
          type: integer
          format: int64
        DegradedInference:
          type: integer
          format: int64
        DegradedInput:
          type: integer
          format: int64
    ReliabilitySummary:
      type: object
      properties:
        StartTime:
          type: string
          format: date-time
        EndTime:
          type: string
          format: date-time
        StreamSuccessRate:
          type: number
          format: double
        NoOrchAvailableRate:
          type: number
          format: double
        OrchSwapCount:
          type: integer
          format: int64
        OrchSwapRate:
          type: number
          format: double
        InferenceRestartRate:
          type: number
          format: double
        DegradedStateRate:
          type: number
          format: double
        FailureBreakdown:
          $ref: "#/components/schemas/FailureBreakdown"
    ReliabilityBucket:
      type: object
      properties:
        Timestamp:
          type: string
          format: date-time
        Started:
          type: integer
          format: int64
        SuccessRate:
          type: number
          format: double
          nullable: true
          description: Null when sample count is below minimum threshold (5 streams).
        NoOrchAvailableRate:
          type: number
          format: double
          nullable: true
        DegradedRate:
          type: number
          format: double
          nullable: true
    OrchReliability:
      type: object
      properties:
        Address:
          type: string
        StreamsHandled:
          type: integer
          format: int64
        DegradedRate:
          type: number
          format: double
        RestartRate:
          type: number
          format: double
        ErrorRate:
          type: number
          format: double
    FailureEvent:
      type: object
      properties:
        Timestamp:
          type: string
          format: date-time
        FailureType:
          type: string
          enum: [no_orch_available, orch_swap, inference_restart, inference_error]
        StreamID:
          type: string
        RequestID:
          type: string
        Org:
          type: string
        Gateway:
          type: string
        Detail:
          type: string
          description: Raw JSON payload from the source event.

    # ── Leaderboard (R6) ────────────────────────────────────────────────────
...
    E2ELatencyStats:
      type: object
      properties:
        AvgMS: { type: number, format: double }
        P50MS: { type: number, format: double }
        P95MS: { type: number, format: double }
        P99MS: { type: number, format: double }
        SampleCount: { type: integer, format: int64 }
    PipelineE2ELatency:
      type: object
      properties:
        Pipeline: { type: string }
        AvgMS: { type: number, format: double }
        P50MS: { type: number, format: double }
        P95MS: { type: number, format: double }
        P99MS: { type: number, format: double }
        SampleCount: { type: integer, format: int64 }
...
    E2ELatencySummary:
      type: object
      properties:
        StartTime: { type: string, format: date-time }
        EndTime: { type: string, format: date-time }
        Overall: { $ref: "#/components/schemas/E2ELatencyStats" }
        ByPipeline:
          type: array
          items: { $ref: "#/components/schemas/PipelineE2ELatency" }
        ByOrchestrator:
          type: array
          items: { $ref: "#/components/schemas/OrchE2ELatency" }
    E2ELatencyBucket:
      type: object
      properties:
        Timestamp: { type: string, format: date-time }
        AvgMS: { type: number, format: double }
        P50MS: { type: number, format: double }
        P95MS: { type: number, format: double }
        SampleCount: { type: integer, format: int64 }

    # ── Pipelines (R10) ─────────────────────────────────────────────────────
    PipelineSummary:
      type: object
      properties:
        Pipeline: { type: string }
        ActiveStreams: { type: integer, format: int64 }
        StartedCount: { type: integer, format: int64 }
        SuccessRate: { type: number, format: double }
        AvgInferenceFPS: { type: number, format: double }
        TotalPaymentsWEI: { $ref: "#/components/schemas/WEI" }
        WarmOrchCount: { type: integer, format: int64 }
        TopOrchAddress: { type: string }
...
    PipelineDetail:
      type: object
      properties:
        Pipeline: { type: string }
        StartTime: { type: string, format: date-time }
        EndTime: { type: string, format: date-time }
        ActiveStreams: { type: integer, format: int64 }
        StartedCount: { type: integer, format: int64 }
        SuccessRate: { type: number, format: double }
        AvgInferenceFPS: { type: number, format: double }
        TotalPaymentsWEI: { $ref: "#/components/schemas/WEI" }
        WarmOrchCount: { type: integer, format: int64 }
        ByModel:
          type: array
          items: { $ref: "#/components/schemas/ModelPipelineStats" }

    # ── Pricing (R11) ───────────────────────────────────────────────────────
    OrchPricingEntry:
      type: object
      properties:
        OrchAddress: { type: string }
        Name: { type: string }
        Pipeline: { type: string }
        ModelID: { type: string }
        PricePerUnit: { type: integer, format: int64 }
        PixelsPerUnit: { type: integer, format: int64 }
        IsWarm: { type: boolean }
...
    OrchPricingProfile:
      type: object
      properties:
        OrchAddress: { type: string }
        Name: { type: string }
        IsActive: { type: boolean }
        Pipelines:
          type: array
          items: { $ref: "#/components/schemas/PipelinePricing" }

    # ── Model Performance (R12) ──────────────────────────────────────────────
...
    FailuresByPipeline:
      type: object
      properties:
        Pipeline: { type: string }
        NoOrchCount: { type: integer, format: int64 }
        InferenceErrorCount: { type: integer, format: int64 }
        OrchSwapCount: { type: integer, format: int64 }
        InferenceRestartCount: { type: integer, format: int64 }
        TotalFailures: { type: integer, format: int64 }
        FailureRate: { type: number, format: double, description: "total failures / streams started (0–1)" }
    FailuresByOrch:
      type: object
      properties:
        OrchAddress: { type: string }
        Name: { type: string }
        InferenceErrors: { type: integer, format: int64 }
        InferenceRestarts: { type: integer, format: int64 }
        TotalFailures: { type: integer, format: int64 }
        StreamsHandled: { type: integer, format: int64 }
        FailureRate: { type: number, format: double }

    # ── Capacity (R15) ───────────────────────────────────────────────────────
...
  responses:
    429:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    500:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
paths:
...
  /v1/perf/fps:
    get:
      tags: [Performance]
      summary: FPS summary
      description: |
        Inference and input FPS percentiles (p5/p50/p99) broken down by pipeline
        and by orchestrator. Samples with fps == 0 are excluded.
        Computed from raw events (not aggregates) for percentile accuracy.
      operationId: getFPSSummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FPSSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/fps/history:
    get:
      tags: [Performance]
      summary: FPS history
      description: Hourly average inference and input FPS for charting. Buckets flagged as `Degraded` when avg FPS drops ≥20% from the prior bucket.
      operationId: listFPSHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FPSBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/latency:
    get:
      tags: [Performance]
      summary: Orchestrator discovery latency
      description: Avg, p50, p95, p99 discovery latency per orchestrator, computed from raw `discovery_results` events.
      operationId: getLatencySummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatencySummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/webrtc:
    get:
      tags: [Performance]
      summary: WebRTC quality
      description: Aggregated WebRTC video/audio jitter, packet loss, and connection quality distribution.
      operationId: getWebRTCQuality
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/stream_id"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebRTCQuality"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Payments ───────────────────────────────────────────────────────────────
...
  /v1/reliability/summary:
    get:
      tags: [Reliability]
      summary: Reliability summary
      description: Network-wide stream success rate, failure rates, and failure breakdown over the window.
      operationId: getReliabilitySummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReliabilitySummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/reliability/history:
    get:
      tags: [Reliability]
      summary: Reliability history
      description: Hourly success and failure rates for charting. Rates are null when the sample count is below the minimum threshold (5 streams).
      operationId: listReliabilityHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReliabilityBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/reliability/orchs:
    get:
      tags: [Reliability]
      summary: Orchestrator reliability
      description: Per-orchestrator degraded, restart, and error rates over the window.
      operationId: listOrchReliability
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrchReliability"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/failures:
    get:
      tags: [Reliability]
      summary: Failure events
      description: Raw failure event log, newest first. Filterable by type.
      operationId: listFailures
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/failure_type"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FailureEvent"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Leaderboard ────────────────────────────────────────────────────────────
...
  /v1/perf/e2e-latency:
    get:
      summary: E2E latency summary
      description: Returns end-to-end latency percentiles from stream status samples.
      operationId: getE2ELatencySummary
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/E2ELatencySummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/perf/e2e-latency/history:
    get:
      summary: E2E latency history
      description: Returns hourly E2E latency buckets (p50, p95, avg).
      operationId: listE2ELatencyHistory
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/E2ELatencyBucket" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Pipelines (R10) ─────────────────────────────────────────────────────
  /v1/pipelines:
    get:
      summary: List pipelines
      description: Returns a cross-cutting summary per pipeline combining streams, FPS, and payment data.
      operationId: listPipelines
      tags: [Pipelines]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PipelineSummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/pipelines/{pipeline}:
    get:
      summary: Pipeline detail
      description: Returns detailed stats for one pipeline including model breakdown.
      operationId: getPipelineDetail
      tags: [Pipelines]
      parameters:
        - name: pipeline
          in: path
          required: true
          description: Pipeline name (e.g. `text-to-image`).
          schema: { type: string }
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Pricing (R11) ───────────────────────────────────────────────────────
  /v1/net/pricing:
    get:
      summary: List orchestrator pricing
      description: Returns flat pricing rows for all active orchestrators.
      operationId: listPricing
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrchPricingEntry" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/pricing/{address}:
    get:
      summary: Orchestrator pricing profile
      description: Returns structured pipeline/model pricing for one orchestrator.
      operationId: getOrchPricingProfile
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchPricingProfile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Model Performance (R12) ──────────────────────────────────────────────
...
  /v1/failures/by-pipeline:
    get:
      summary: Failures by pipeline
      description: Returns aggregated failure counts per pipeline.
      operationId: listFailuresByPipeline
      tags: [Reliability]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FailuresByPipeline" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/failures/by-orch:
    get:
      summary: Failures by orchestrator
      description: Returns aggregated inference error and restart counts per orchestrator.
      operationId: listFailuresByOrch
      tags: [Reliability]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/FailuresByOrch" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Capacity (R15) ───────────────────────────────────────────────────────
...